plantfeed-backend-mobile/requirements.txt
appdirs==1.4.4
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.0
Django==3.2.9
django-filter==22.1
djangorestframework==3.14.0
openai==0.27.8
djangorestframework-simplejwt==5.2.2
fernet==1.0.1
idna==3.4
Markdown==3.4.1
mysqlclient==2.1.1
pyaes==1.6.1
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
django-allauth
Pillow==9.4.0
django-rest-auth
djoser
gunicorn
django-heroku
stripe
django-environ
django-mathfilters